\name{dist2cluster}
\alias{dist2cluster}

\title{
 cluster with distance matrix by igraph clustering algorithms
}

\description{
 cluster with distance matrix by igraph clustering algorithms
}

\usage{
dist2cluster(dat.d=dat.d, wt, method="louvain");
}

\arguments{
   \item{dat.d}{distance matrix or object}
   \item{wt}{weight or cutoff}
   \item{method}{should be one of louvain, fast_greedy, infomap,
                  label_prop, spinglass and weight}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

