\name{dist2graph}
\alias{dist2graph}

\title{
 create graph by distance matrix
}

\description{
 create graph by distance matrix
}

\usage{
  dist2graph(dat.d, w=2);
}

\arguments{
   \item{dat.d}{distance object}
   \item{w}{weight or cutoff}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

