\name{dist2tsne}
\alias{dist2tsne}

\title{
 do tSNE with distance matrix
}

\description{
 do tSNE with distance matrix
}

\usage{
  dist2tsne(dat.d, tsne.per, tsne.max);
}

\arguments{
   \item{dat.d}{distance object}
   \item{tsne.per}{perplexity of tsne}
   \item{tsne.max}{perplexity of max_iter}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

