\name{do.scale}
\alias{do.scale}

\title{
 local scale function
}

\description{
 local scale function
}

\usage{
  do.scale(x, newMax, newMin)
}

\arguments{
   \item{x}{numerical vector}
   \item{newMax}{maximum of scale}
   \item{newMin}{minimum of scale}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

