\name{do_filter}
\alias{do_filter}

\title{
 remove rows by the genes
}

\description{
 remove rows by the genes
}

\usage{
  do_filter(dat, g);
}

\arguments{
   \item{dat}{data.frame or matrix: column 1 and 2 are the genes from association analysis.}
   \item{g}{genes that will be excluded.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

