\name{l4xy}
\alias{l4xy}

\title{
 map coordinates on the given circle 
}

\description{
 map coordinates on the given circle 
}

\usage{
  l4xy(l, xy, r)
}

\arguments{
   \item{l}{numerical vector}
   \item{xy}{coordinate}
   \item{r}{radius}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

