\name{p2dist}
\alias{p2dist}

\title{
 do distance matrix by p values
}

\description{
 do distance matrix by p values
}

\usage{
  p2dist(dat=dat, p=p);
}

\arguments{
   \item{dat}{data.frame or matrix: column 1 and 2 are the genes from association analysis.}
   \item{p}{the p values from the association analysis}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

