\name{pathway_alignment}
\alias{pathway_alignment}

\title{
  align mutual exclusivity mutation module and pathway network
}

\description{
 align mutual exclusivity mutation module and pathway network
}

\usage{
  pathway_alignment(g0, relations)
}

\arguments{
   \item{g0}{mutual exclusivity mutation module (graph object)}
   \item{relations}{data frame: column 1 is for gene 1, column 2 is for gene 2.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

