\name{pathway_map}
\alias{pathway_map}

\title{
 map genes on pathway
}

\description{
 map genes on pathway
}

\usage{
  pathway_map(genes, relations)
}

\arguments{
   \item{genes}{a vector of genes}
   \item{relations}{data frame: column 1 is for gene 1, column 2 is for gene 2.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

