\name{plot_me}
\alias{plot_me}

\title{
 visulization of mutual exclusivity
}

\description{
 visulization of mutual exclusivity
}

\usage{
  plot_me(x, sort=T, s.lab="", main="", g.cex=1, g.col="black",
          m.cex=1, s.cex=1, col.bg="white", col="blue", horiz=T)
}

\arguments{
   \item{x}{data.frame or matrix of mutation}
   \item{sort}{logical: if sorting by mutation number}
   \item{s.lab}{mutation gene label}
   \item{main}{main}
   \item{g.cex}{cex for gene label}
   \item{g.col}{color for gene label}
   \item{m.cex}{cex for mutation}
   \item{s.cex}{cex for sample}
   \item{col.bg}{background color}
   \item{col}{mutation color}
   \item{horiz}{logical: horizontal plots, FALSE will be vertical}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

