\name{plot_network}
\alias{plot_network}

\title{
 plot a nice network
}

\description{
 plot a nice network
}

\usage{
  plot_network (graph, label.gene, label.col=NULL, main=NULL)
}

\arguments{
   \item{graph}{graph object}
   \item{label.gene}{genes will be labeled.}
   \item{label.col}{colors are for the labeled genes.}
   \item{main}{main of the plot}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

